Move actual execution out of headcrab_inject
